WIP: Player creates but collision system has corrupted 64-bit pointers
Progress:
- ALINK player now creates (face BCK failure made non-fatal)
- Scene Execute waits for player creation
- CrrPos crash protection in ALINK create
- BGW pointer validation in d_bg_s.cpp collision checks

Root issue still open: cBgS_ChkElm.m_bgw_base_ptr contains
0x1018800000018 (corrupted 64-bit pointer). This is likely a
32→64 bit struct size mismatch in the collision element registration
path. The collision data loader stores pointers that are only
32 bits wide into fields that are now 64 bits on PC.

Next: investigate cBgS::RegistBgw / cBgS_ChkElm initialization
to find where the pointer corruption originates.

+++ pc/src/pc_aram.cpp
@@ -28,7 +28,7 @@ u32 ARGetSize(void) { return PC_ARAM_SIZE; }
 u32 ARGetBaseAddress(void) { return 0; }
 u32 ARGetInternalSize(void) { return PC_ARAM_SIZE; }
 
-void ARStartDMA(u32 type, u32 mainmem_addr, u32 aram_addr, u32 length) {
+void ARStartDMA(u32 type, uintptr_t mainmem_addr, u32 aram_addr, u32 length) {
     (void)type; (void)mainmem_addr; (void)aram_addr; (void)length;
     /* Will implement ARAM<->main memory DMA when needed */
 }
@@ -41,7 +41,7 @@ int ARGetDMAStatus(void) { return 0; /* idle */ }
 void ARQInit(void) {}
 void ARQReset(void) {}
 void ARQPostRequest(void* task, u32 owner, u32 type, u32 priority,
-                    u32 source, u32 dest, u32 length, void* callback) {
+                    uintptr_t source, uintptr_t dest, u32 length, void* callback) {
     (void)task; (void)owner; (void)type; (void)priority;
     (void)source; (void)dest; (void)length; (void)callback;
 }
